+++ Assets/Prefabs/Gameplay Systems/RespawnManager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fbe6e6d098ae2184ebf72d50cae03ecf
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
