--- Assets/Scripts/Enemy.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a5516659beb4b4f87a7669cd40f72e22
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
