+++ Assets/Scripts/Enemy/WallPoint.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f8f7586aad127054da561da535cd8781
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
