--- Assets/Scripts/OrangeObjectManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5f2219be4ddbd644ca840958b0c3200b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
