+++ Assets/Scripts/PlayerMovement.cs
@@ -27,8 +27,7 @@ public class PlayerMovement : MonoBehaviour
     private AudioSource playerSound;
     // State variables
 
-    [SerializeField] public ScribbleWall scribbleWall;
-    [SerializeField] public OrangeObjectManager orangeObjectManager;
+    public RespawnManager respawner;
     public Vector3 respawnPos;
     private bool isDead = false;
     [SerializeField] private bool isPaused = false;
@@ -46,6 +45,7 @@ public class PlayerMovement : MonoBehaviour
         playerSound = GetComponent<AudioSource>();
         audioSys = FindObjectOfType<AudioSystem>();
 
+        respawner = FindObjectOfType<RespawnManager>();
         // Set initial respawn position
         respawnPos = transform.position;
         if (animateSpawnOnLoad)
@@ -183,10 +183,7 @@ public class PlayerMovement : MonoBehaviour
         rb2d.simulated = true;
         rb2d.velocity = Vector2.zero;
 
-        // Respawn orange objects and scribble wall if they exist
-        scribbleWall?.StartRespawn();
-        orangeObjectManager?.StartOrangeObjectsRespawn();
-
+        respawner.StartObjectRespawn();    
         // Move the player to the respawn position
         transform.position = respawnPos;
