+++ Assets/Scripts/RespawnInterface.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public abstract class RespawnInterface : MonoBehaviour
+
+{
+    public abstract void StartRespawn();
+}
