+++ Assets/Scripts/RespawnInterface.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 665c97c31cfc9a040a4bab4c5acf20b7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
