+++ Assets/Scripts/RespawnManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 361cb2d8b591c4f4985cf4d851d95fcb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
