--- Assets/Scripts/ScribbleWall.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c27091dbcd24dd84988b82ab6287e79f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
