+++ Assets/Sprites/Characters/Mike/Mike Save Stickman.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 046d5fc8821761348a06f1c7a274e6b0
+guid: 2ac9c06a5e46ecd40ad454c7db931983
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
